change back to long

+++ activity-microservice/src/main/java/nl/tudelft/sem/template/example/controllers/ActivityController.java
@@ -39,7 +39,7 @@ public class ActivityController {
 
 
     @PutMapping("/edit/{id}")
-    public void editActivity(@RequestParam(value = "id", required = true) @PathVariable int id, @ModelAttribute("request") @RequestBody ActivityRequestModel request) throws UnauthorizedException, ActivityNotFoundException {
+    public void editActivity(@RequestParam(value = "id", required = true) @PathVariable long id, @ModelAttribute("request") @RequestBody ActivityRequestModel request) throws UnauthorizedException, ActivityNotFoundException {
         NetId username = new NetId(authManager.getNetId());
         activityService.editActivity(username, id, request);
     }
@@ -77,7 +77,7 @@ public class ActivityController {
     }
 
     @GetMapping("/user/{id}")
-    public NetId getOwnerById(@PathVariable("id") int id) throws ActivityNotFoundException {
+    public NetId getOwnerById(@PathVariable("id") long id) throws ActivityNotFoundException {
         return activityService.getById(id).getOwner();
     }
 
@@ -87,7 +87,7 @@ public class ActivityController {
     }
 
     @GetMapping("/activityId/{id}")
-    public Activity getById(@PathVariable("id") int id) throws ActivityNotFoundException {
+    public Activity getById(@PathVariable("id") long id) throws ActivityNotFoundException {
         return activityService.getById(id);
     }
 }
